+++ ftag/hdf5/__init__.py
@@ -1,5 +1,17 @@
 from ftag.hdf5.h5reader import H5Reader
-from ftag.hdf5.h5utils import cast_dtype, get_dtype, get_dummy_file
+from ftag.hdf5.h5utils import (
+    cast_dtype,
+    get_dtype,
+    get_dummy_file,
+    join_structured_arrays,
+)
 from ftag.hdf5.h5writer import H5Writer
 
-__all__ = ["H5Reader", "H5Writer", "get_dummy_file", "get_dtype", "cast_dtype"]
+__all__ = [
+    "H5Reader",
+    "H5Writer",
+    "get_dummy_file",
+    "get_dtype",
+    "cast_dtype",
+    "join_structured_arrays",
+]