+++ ftag/hdf5/h5utils.py
@@ -1,10 +1,10 @@
-from tempfile import NamedTemporaryFile
+from tempfile import NamedTemporaryFile, mkdtemp
 
 import h5py
 import numpy as np
 from numpy.lib.recfunctions import unstructured_to_structured as u2s
 
-__all__ = ["get_dummy_file"]
+__all__ = ["get_dummy_file", "join_structured_arrays"]
 
 
 def get_dtype(ds, variables: list[str] | None = None, precision: str | None = None) -> np.dtype:
@@ -118,7 +118,7 @@ def get_dummy_file():
     valid = valid.astype(bool).view(dtype=np.dtype([("valid", bool)]))
     tracks = join_structured_arrays([tracks, valid])
 
-    fname = NamedTemporaryFile(suffix=".h5").name
+    fname = NamedTemporaryFile(suffix=".h5", dir=mkdtemp()).name
     f = h5py.File(fname, "w")
     f.create_dataset("jets", data=jets)
     f.create_dataset("tracks", data=tracks)