+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### [Latest]
 
+- add function to estimate number of available jets [#6](https://github.com/umami-hep/atlas-ftag-tools/pull/6)
+
 ### [v0.0.2]
 
 fixed version of buggy release